test(agents): move expanded context coverage to its sibling module
Keep one home for the context tests now that sub-cap regressions take the inline module beyond the 500-line invariant.
